refactor(xmp): consolidate parser implementations
